[FEATURE] separate FeatureXML and FeatureXMLHandler
