+++ src/openms/include/OpenMS/FORMAT/FeatureXMLFile.h
@@ -36,13 +36,6 @@
 
 #include <OpenMS/FORMAT/OPTIONS/FeatureFileOptions.h>
 #include <OpenMS/FORMAT/XMLFile.h>
-#include <OpenMS/FORMAT/HANDLERS/XMLHandler.h>
-#include <OpenMS/CONCEPT/ProgressLogger.h>
-#include <OpenMS/METADATA/ProteinIdentification.h>
-#include <OpenMS/METADATA/PeptideIdentification.h>
-#include <OpenMS/DATASTRUCTURES/ConvexHull2D.h>
-#include <OpenMS/DATASTRUCTURES/Param.h>
-#include <OpenMS/DATASTRUCTURES/Map.h>
 
 #include <iosfwd>
 
@@ -66,9 +59,7 @@ namespace OpenMS
     @ingroup FileIO
   */
   class OPENMS_DLLAPI FeatureXMLFile :
-    protected Internal::XMLHandler,
-    public Internal::XMLFile,
-    public ProgressLogger
+    public Internal::XMLFile
   {
 
 public:
@@ -109,85 +100,8 @@ public:
 
 protected:
 
-    // restore default state for next load/store operation
-    void resetMembers_();
-
-    // Docu in base class
-    void endElement(const XMLCh* const /*uri*/, const XMLCh* const /*local_name*/, const XMLCh* const qname) override;
-
-    // Docu in base class
-    void startElement(const XMLCh* const /*uri*/, const XMLCh* const /*local_name*/, const XMLCh* const qname, const xercesc::Attributes& attributes) override;
-
-    // Docu in base class
-    void characters(const XMLCh* const chars, const XMLSize_t length) override;
-
-    /// Writes a feature to a stream
-    void writeFeature_(const String& filename, std::ostream& os, const Feature& feat, const String& identifier_prefix, UInt64 identifier, UInt indentation_level);
-
-    /// Writes a peptide identification to a stream (for assigned/unassigned peptide identifications)
-    void writePeptideIdentification_(const String& filename, std::ostream& os, const PeptideIdentification& id, const String& tag_name, UInt indentation_level);
-
-
-    /**
-        @brief update the pointer to the current feature
-
-        @param create If true, a new (empty) Feature is added at the appropriate subordinate_feature_level_
-    */
-    void updateCurrentFeature_(bool create);
-
-    /// allows for early return in parsing functions when certain sections should be ignored
-    /// <=0 - parsing ON
-    ///  >0 - this number of tags have been entered that forbid parsing and need to be exited before parsing continues
-    Int disable_parsing_;
-
-    /// points to the last open &lt;feature&gt; tag (possibly a subordinate feature)
-    Feature* current_feature_;
-    /// Feature map pointer for reading
-    FeatureMap* map_;
     /// Options that can be set
     FeatureFileOptions options_;
-    /// only parse until "count" tag is reached (used in loadSize())
-    bool size_only_;
-    /// holds the putative size given in count
-    Size expected_size_;
-
-    /**@name temporary data structures to hold parsed data */
-    //@{
-    Param param_;
-    ConvexHull2D::PointArrayType current_chull_;
-    DPosition<2> hull_position_;
-    //@}
-
-    /// current dimension of the feature position, quality, or convex hull point
-    UInt dim_;
-
-    /// for downward compatibility, all tags in the old description must be ignored
-    bool in_description_;
-
-    /// level in Feature stack during parsing
-    Int subordinate_feature_level_;
-
-    /// Pointer to last read object as a MetaInfoInterface, or null.
-    MetaInfoInterface* last_meta_;
-
-    /// Temporary protein ProteinIdentification
-    ProteinIdentification prot_id_;
-    /// Temporary peptide ProteinIdentification
-    PeptideIdentification pep_id_;
-    /// Temporary protein hit
-    ProteinHit prot_hit_;
-    /// Temporary peptide hit
-    PeptideHit pep_hit_;
-    /// Map from protein id to accession
-    Map<String, String> proteinid_to_accession_;
-    /// Map from search identifier concatenated with protein accession to id
-    Map<String, Size> accession_to_id_;
-    /// Map from identification run identifier to file xs:id (for linking peptide identifications to the corresponding run)
-    Map<String, String> identifier_id_;
-    /// Map from file xs:id to identification run identifier (for linking peptide identifications to the corresponding run)
-    Map<String, String> id_identifier_;
-    /// Temporary search parameters file
-    ProteinIdentification::SearchParameters search_param_;
 
   };
