+++ src/openms/include/OpenMS/FORMAT/IdXMLFile.h
@@ -44,6 +44,10 @@
 
 namespace OpenMS
 {
+  namespace Internal
+  {
+    class FeatureXMLHandler;
+  }
   /**
     @brief Used to load and store idXML files
 
@@ -68,7 +72,7 @@ namespace OpenMS
 public:
     // both ConsensusXMLFile and FeatureXMLFile use some protected IdXML helper functions to parse identifications without code duplication
     friend class ConsensusXMLFile;
-    friend class FeatureXMLFile;
+    friend class Internal::FeatureXMLHandler;
 
     /// Constructor
     IdXMLFile();
